alias reworking

+++ scripts/functions/fuck.fish
@@ -1,4 +1,7 @@
 function fuck
-	git reset --hard $argv
-
+	if [ (count $argv) -gt 1 ]
+		git checkout $argv
+	else
+		git reset --hard $argv
+	end
 end
